backend/requirements.txt
fastapi[standard]
uvicorn
sqlmodel
alembic
asyncpg
pydantic-settings
pydantic-settings
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-jose[cryptography]
python-multipart
email-validator
pytest
pytest-asyncio
httpx
aiosqlite
greenlet

